feat: create uploader for json file and show its contents
Create a component HomePage to show the title and subtitle of the
application in order to inform how is its purpose, also, this same
component is responsible to handle the json file load and save its
content in a variable. After handle the json file content, emits
an event to parent component with file name and content.

In App.vue component, when file content is received by the HomePage
component, save its content in a ref variable and change the content
of the screen to show the json text, to change the content is used
a v-if directive to hide content of the HomePage when there's a file
loaded.