feat: use virtual scroll to render a json with many lines
The JsonTreeItem.vue component was deleted because of the new approach
to render the json tree.

Now, instead load in blocks of types (array or objects), the JsonTree.vue
component is responsible to compute all lines to be rendered,
based on the quantity of the lines to render, it will compute a
script to create a virtual scroll list.