feat: add button to upload new json file
Add new button on json tree if user wants to upload a new file
without gives a refresh on page.

Also did a small refactoring in button css styles, to reuse the
styles already applied on home page in this new button in json tree,
it moved the styles for global css file and rename the class in
home page component.